refactor(codec): internal restructure of codec traits
No changes, just code movement
