+++ zarrs_codec/src/codec_traits/array_partial_async.rs
@@ -1,0 +1,110 @@
+use std::any::Any;
+use std::sync::Arc;
+
+use zarrs_chunk_grid::Indexer;
+use zarrs_data_type::DataType;
+use zarrs_plugin::{MaybeSend, MaybeSync};
+use zarrs_storage::StorageError;
+
+use crate::{
+    ArrayBytes, ArrayBytesDecodeIntoTarget, CodecError, CodecOptions, InvalidNumberOfElementsError,
+    decode_into_array_bytes_target,
+};
+
+/// Asynchronous partial array decoder traits.
+#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
+#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+pub trait AsyncArrayPartialDecoderTraits: Any + MaybeSend + MaybeSync {
+    /// Return the data type of the partial decoder.
+    fn data_type(&self) -> &DataType;
+
+    /// Returns whether the chunk exists.
+    ///
+    /// # Errors
+    /// Returns [`StorageError`] if a storage operation fails.
+    async fn exists(&self) -> Result<bool, StorageError>;
+
+    /// Returns the size of chunk bytes held by the partial decoder.
+    ///
+    /// Intended for use by size-constrained partial decoder caches.
+    fn size_held(&self) -> usize;
+
+    /// Partially decode a chunk.
+    ///
+    /// # Errors
+    /// Returns [`CodecError`] if a codec fails, array subset is invalid, or the array subset shape does not match array view subset shape.
+    async fn partial_decode<'a>(
+        &'a self,
+        indexer: &dyn Indexer,
+        options: &CodecOptions,
+    ) -> Result<ArrayBytes<'a>, CodecError>;
+
+    /// Partially decode into a preallocated output.
+    ///
+    /// This method is intended for internal use by Array.
+    /// It currently only works for fixed length data types.
+    ///
+    /// The `indexer` shape and dimensionality does not need to match `output_subset`, but the number of elements must match.
+    /// Extracted elements from the `indexer` are written as ordered by the indexer.
+    /// For an [`ArraySubset`](zarrs_chunk_grid::ArraySubset), that is C order.
+    ///
+    /// # Errors
+    /// Returns [`CodecError`] if a codec fails or the number of elements in `indexer` does not match the number of elements in `output_view`,
+    #[allow(clippy::missing_safety_doc)]
+    async fn partial_decode_into(
+        &self,
+        indexer: &dyn Indexer,
+        output_target: ArrayBytesDecodeIntoTarget<'_>,
+        options: &CodecOptions,
+    ) -> Result<(), CodecError> {
+        if indexer.len() != output_target.num_elements() {
+            return Err(InvalidNumberOfElementsError::new(
+                indexer.len(),
+                output_target.num_elements(),
+            )
+            .into());
+        }
+
+        let decoded_value = self.partial_decode(indexer, options).await?;
+        decode_into_array_bytes_target(&decoded_value, output_target)
+    }
+
+    /// Returns whether this decoder supports partial decoding.
+    ///
+    /// If this returns `true`, the decoder can efficiently handle partial decoding operations.
+    /// If this returns `false`, partial decoding will fall back to a full decode operation.
+    fn supports_partial_decode(&self) -> bool;
+}
+
+/// Asynchronous partial array encoder traits.
+#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
+#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+pub trait AsyncArrayPartialEncoderTraits:
+    AsyncArrayPartialDecoderTraits + Any + MaybeSend + MaybeSync
+{
+    /// Return the encoder as an [`Arc<AsyncArrayPartialDecoderTraits>`].
+    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn AsyncArrayPartialDecoderTraits>;
+
+    /// Erase the chunk.
+    ///
+    /// # Errors
+    /// Returns an error if there is an underlying store error.
+    async fn erase(&self) -> Result<(), CodecError>;
+
+    /// Partially encode a chunk.
+    ///
+    /// # Errors
+    /// Returns [`CodecError`] if a codec fails or an array subset is invalid.
+    async fn partial_encode(
+        &self,
+        indexer: &dyn Indexer,
+        bytes: &ArrayBytes<'_>,
+        options: &CodecOptions,
+    ) -> Result<(), CodecError>;
+
+    /// Returns whether this encoder supports partial encoding.
+    ///
+    /// If this returns `true`, the encoder can efficiently handle partial encoding operations.
+    /// If this returns `false`, partial encoding will fall back to a full decode and encode operation.
+    fn supports_partial_encode(&self) -> bool;
+}
