+++ zarrs_codec/src/codec_traits/array_partial_sync.rs
@@ -1,0 +1,107 @@
+use std::any::Any;
+use std::sync::Arc;
+
+use zarrs_chunk_grid::Indexer;
+use zarrs_data_type::DataType;
+use zarrs_plugin::{MaybeSend, MaybeSync};
+use zarrs_storage::StorageError;
+
+use crate::{
+    ArrayBytes, ArrayBytesDecodeIntoTarget, CodecError, CodecOptions, InvalidNumberOfElementsError,
+    decode_into_array_bytes_target,
+};
+
+/// Partial array decoder traits.
+pub trait ArrayPartialDecoderTraits: Any + MaybeSend + MaybeSync {
+    /// Return the data type of the partial decoder.
+    fn data_type(&self) -> &DataType;
+
+    /// Returns whether the chunk exists.
+    ///
+    /// # Errors
+    /// Returns [`StorageError`] if a storage operation fails.
+    fn exists(&self) -> Result<bool, StorageError>;
+
+    /// Returns the size of chunk bytes held by the partial decoder.
+    ///
+    /// Intended for use by size-constrained partial decoder caches.
+    fn size_held(&self) -> usize;
+
+    /// Partially decode a chunk.
+    ///
+    /// If the inner `input_handle` is a bytes decoder and partial decoding returns [`None`], then the array subsets have the fill value.
+    ///
+    /// # Errors
+    /// Returns [`CodecError`] if a codec fails or an array subset is invalid.
+    fn partial_decode(
+        &self,
+        indexer: &dyn Indexer,
+        options: &CodecOptions,
+    ) -> Result<ArrayBytes<'_>, CodecError>;
+
+    /// Partially decode into a preallocated output.
+    ///
+    /// This method is intended for internal use by Array.
+    /// It currently only works for fixed length data types.
+    ///
+    /// The `indexer` shape and dimensionality does not need to match `output_subset`, but the number of elements must match.
+    /// Extracted elements from the `indexer` are written as ordered by the indexer.
+    /// For an [`ArraySubset`](zarrs_chunk_grid::ArraySubset), that is C order.
+    ///
+    /// # Errors
+    /// Returns [`CodecError`] if a codec fails or the number of elements in `indexer` does not match the number of elements in `output_view`,
+    fn partial_decode_into(
+        &self,
+        indexer: &dyn Indexer,
+        output_target: ArrayBytesDecodeIntoTarget<'_>,
+        options: &CodecOptions,
+    ) -> Result<(), CodecError> {
+        if indexer.len() != output_target.num_elements() {
+            return Err(InvalidNumberOfElementsError::new(
+                indexer.len(),
+                output_target.num_elements(),
+            )
+            .into());
+        }
+
+        let decoded_value = self.partial_decode(indexer, options)?;
+        decode_into_array_bytes_target(&decoded_value, output_target)
+    }
+
+    /// Returns whether this decoder supports partial decoding.
+    ///
+    /// If this returns `true`, the decoder can efficiently handle partial decoding operations.
+    /// If this returns `false`, partial decoding will fall back to a full decode operation.
+    fn supports_partial_decode(&self) -> bool;
+}
+
+/// Partial array encoder traits.
+pub trait ArrayPartialEncoderTraits:
+    ArrayPartialDecoderTraits + Any + MaybeSend + MaybeSync
+{
+    /// Return the encoder as an [`Arc<ArrayPartialDecoderTraits>`].
+    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn ArrayPartialDecoderTraits>;
+
+    /// Erase the chunk.
+    ///
+    /// # Errors
+    /// Returns an error if there is an underlying store error.
+    fn erase(&self) -> Result<(), CodecError>;
+
+    /// Partially encode a chunk.
+    ///
+    /// # Errors
+    /// Returns [`CodecError`] if a codec fails or an array subset is invalid.
+    fn partial_encode(
+        &self,
+        indexer: &dyn Indexer,
+        bytes: &ArrayBytes<'_>,
+        options: &CodecOptions,
+    ) -> Result<(), CodecError>;
+
+    /// Returns whether this encoder supports partial encoding.
+    ///
+    /// If this returns `true`, the encoder can efficiently handle partial encoding operations.
+    /// If this returns `false`, partial encoding will fall back to a full decode and encode operation.
+    fn supports_partial_encode(&self) -> bool;
+}
