+++ zarrs_codec/src/lib.rs
@@ -12,15 +12,30 @@ mod options;
 pub use codec_specific_options::CodecSpecificOptions;
 
 mod codec_partial_default;
+use codec_partial_default::BytesToBytesCodecPartialDefault;
 pub use codec_partial_default::CodecPartialDefault;
-use codec_partial_default::{
-    ArrayToArrayCodecPartialDefault, ArrayToBytesCodecPartialDefault,
-    BytesToBytesCodecPartialDefault,
-};
 
 mod array_bytes_fixed_disjoint_view;
 pub use array_bytes_fixed_disjoint_view::{
     ArrayBytesFixedDisjointView, ArrayBytesFixedDisjointViewCreateError,
+};
+
+mod codec_traits;
+pub use codec_traits::array::ArrayCodecTraits;
+pub use codec_traits::array_partial_sync::{ArrayPartialDecoderTraits, ArrayPartialEncoderTraits};
+pub use codec_traits::array_to_array::{ArrayToArrayCodecTraits, UnboundArrayToArrayCodecTraits};
+pub use codec_traits::array_to_bytes::{ArrayToBytesCodecTraits, UnboundArrayToBytesCodecTraits};
+pub use codec_traits::bytes_partial_sync::{BytesPartialDecoderTraits, BytesPartialEncoderTraits};
+pub use codec_traits::bytes_to_bytes::BytesToBytesCodecTraits;
+pub use codec_traits::{CodecTraits, CodecTraitsV2, CodecTraitsV3};
+
+#[cfg(feature = "async")]
+pub use codec_traits::array_partial_async::{
+    AsyncArrayPartialDecoderTraits, AsyncArrayPartialEncoderTraits,
+};
+#[cfg(feature = "async")]
+pub use codec_traits::bytes_partial_async::{
+    AsyncBytesPartialDecoderTraits, AsyncBytesPartialEncoderTraits,
 };
 
 mod recommended_concurrency;
@@ -47,18 +62,14 @@ use derive_more::derive::Display;
 pub use options::{CodecMetadataOptions, CodecOptions};
 use thiserror::Error;
 use zarrs_metadata::{ArrayShape, ChunkShape, Configuration};
-use zarrs_storage::byte_range::{
-    ByteRange, ByteRangeIterator, InvalidByteRangeError, extract_byte_ranges,
-};
+use zarrs_storage::byte_range::{ByteRangeIterator, InvalidByteRangeError, extract_byte_ranges};
 #[cfg(feature = "async")]
 use zarrs_storage::{AsyncReadableStorage, AsyncReadableWritableStorage};
 use zarrs_storage::{
     OffsetBytesIterator, ReadableStorage, ReadableWritableStorage, StorageError, StoreKey,
 };
 
-use std::any::Any;
 use std::borrow::Cow;
-use std::num::NonZeroU64;
 use std::sync::{Arc, LazyLock, Mutex};
 
 use zarrs_chunk_grid::{
@@ -68,9 +79,8 @@ use zarrs_data_type::{DataType, DataTypeFillValueError, FillValue};
 use zarrs_metadata::v2::MetadataV2;
 use zarrs_metadata::v3::MetadataV3;
 use zarrs_plugin::{
-    ExtensionAliases, ExtensionName, MaybeSend, MaybeSync, Plugin, PluginCreateError,
-    PluginUnsupportedError, RuntimePlugin, RuntimeRegistry, ZarrVersion, ZarrVersion2,
-    ZarrVersion3,
+    ExtensionAliases, ExtensionName, Plugin, PluginCreateError, PluginUnsupportedError,
+    RuntimePlugin, RuntimeRegistry, ZarrVersion, ZarrVersion2, ZarrVersion3,
 };
 
 /// A target for decoding array bytes into a preallocated output view.
@@ -370,480 +380,6 @@ impl ExtensionName for Codec {
     }
 }
 
-/// Trait for creating a codec from Zarr V2 metadata.
-///
-/// Types implementing this trait can be registered as V2 codec plugins via [`CodecPluginV2`].
-pub trait CodecTraitsV2 {
-    /// Create a codec from Zarr V2 metadata.
-    ///
-    /// # Errors
-    /// Returns [`CodecCreateError`] if the codec cannot be created.
-    fn create(metadata: &MetadataV2) -> Result<Codec, CodecCreateError>
-    where
-        Self: Sized;
-}
-
-/// Trait for creating a codec from Zarr V3 metadata.
-///
-/// Types implementing this trait can be registered as V3 codec plugins via [`CodecPluginV3`].
-pub trait CodecTraitsV3 {
-    /// Create a codec from Zarr V3 metadata.
-    ///
-    /// # Errors
-    /// Returns [`CodecCreateError`] if the codec cannot be created.
-    fn create(metadata: &MetadataV3) -> Result<Codec, CodecCreateError>
-    where
-        Self: Sized;
-}
-
-/// Codec traits.
-pub trait CodecTraits: ExtensionName + MaybeSend + MaybeSync {
-    /// Create the codec configuration.
-    ///
-    /// A hidden codec (e.g. a cache) will return [`None`], since it will not have any associated metadata.
-    fn configuration(
-        &self,
-        version: ZarrVersion,
-        options: &CodecMetadataOptions,
-    ) -> Option<Configuration>;
-
-    /// Create the Zarr V3 codec configuration.
-    fn configuration_v3(&self, options: &CodecMetadataOptions) -> Option<Configuration> {
-        self.configuration(ZarrVersion::V3, options)
-    }
-
-    /// Create the Zarr V2 codec configuration.
-    fn configuration_v2(&self, options: &CodecMetadataOptions) -> Option<Configuration> {
-        self.configuration(ZarrVersion::V2, options)
-    }
-
-    /// Returns the partial decoder capability of this codec.
-    fn partial_decoder_capability(&self) -> PartialDecoderCapability;
-
-    /// Returns the partial encoder capability of this codec.
-    fn partial_encoder_capability(&self) -> PartialEncoderCapability;
-}
-
-/// Traits shared by context-bound array codecs.
-pub trait ArrayCodecTraits: MaybeSend + MaybeSync {
-    /// Returns this codec as [`Any`].
-    fn as_any(&self) -> &dyn Any;
-
-    /// Return the decoded data type bound to this codec.
-    fn data_type(&self) -> &DataType;
-
-    /// Return the decoded fill value bound to this codec.
-    fn fill_value(&self) -> &FillValue;
-
-    /// Return the recommended concurrency for the requested decoded shape.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if the decoded `shape` is not valid for the codec.
-    fn recommended_concurrency(
-        &self,
-        shape: &[NonZeroU64],
-    ) -> Result<RecommendedConcurrency, CodecError>;
-}
-
-/// Partial bytes decoder traits.
-pub trait BytesPartialDecoderTraits: Any + MaybeSend + MaybeSync {
-    /// Returns whether the chunk exists.
-    ///
-    /// # Errors
-    /// Returns [`StorageError`] if a storage operation fails.
-    fn exists(&self) -> Result<bool, StorageError>;
-
-    /// Returns the size of chunk bytes held by the partial decoder.
-    ///
-    /// Intended for use by size-constrained partial decoder caches.
-    fn size_held(&self) -> usize;
-
-    /// Partially decode a byte range.
-    ///
-    /// Returns [`None`] if partial decoding of the input handle returns [`None`].
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or the byte range is invalid.
-    fn partial_decode(
-        &self,
-        decoded_region: ByteRange,
-        options: &CodecOptions,
-    ) -> Result<Option<ArrayBytesRaw<'_>>, CodecError> {
-        Ok(self
-            .partial_decode_many(Box::new([decoded_region].into_iter()), options)?
-            .map(|mut v| v.pop().expect("single byte range")))
-    }
-
-    /// Partially decode byte ranges.
-    ///
-    /// Returns [`None`] if partial decoding of the input handle returns [`None`].
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or a byte range is invalid.
-    fn partial_decode_many(
-        &self,
-        decoded_regions: ByteRangeIterator,
-        options: &CodecOptions,
-    ) -> Result<Option<Vec<ArrayBytesRaw<'_>>>, CodecError>;
-
-    /// Decode all bytes.
-    ///
-    /// Returns [`None`] if partial decoding of the input handle returns [`None`].
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails.
-    fn decode(&self, options: &CodecOptions) -> Result<Option<ArrayBytesRaw<'_>>, CodecError> {
-        self.partial_decode(ByteRange::FromStart(0, None), options)
-    }
-
-    /// Returns whether this decoder supports partial decoding.
-    ///
-    /// If this returns `true`, the decoder can efficiently handle partial decoding operations.
-    /// If this returns `false`, partial decoding will fall back to a full decode operation.
-    fn supports_partial_decode(&self) -> bool;
-}
-
-#[cfg(feature = "async")]
-/// Asynchronous partial bytes decoder traits.
-#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
-pub trait AsyncBytesPartialDecoderTraits: Any + MaybeSend + MaybeSync {
-    /// Returns whether the chunk exists.
-    ///
-    /// # Errors
-    /// Returns [`StorageError`] if a storage operation fails.
-    async fn exists(&self) -> Result<bool, StorageError>;
-
-    /// Returns the size of chunk bytes held by the partial decoder.
-    ///
-    /// Intended for use by size-constrained partial decoder caches.
-    fn size_held(&self) -> usize;
-
-    /// Partially decode a byte range.
-    ///
-    /// Returns [`None`] if partial decoding of the input handle returns [`None`].
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or the byte range is invalid.
-    async fn partial_decode<'a>(
-        &'a self,
-        decoded_region: ByteRange,
-        options: &CodecOptions,
-    ) -> Result<Option<ArrayBytesRaw<'a>>, CodecError> {
-        Ok(self
-            .partial_decode_many(Box::new([decoded_region].into_iter()), options)
-            .await?
-            .map(|mut v| v.pop().expect("single byte range")))
-    }
-
-    /// Partially decode byte ranges.
-    ///
-    /// Returns [`None`] if partial decoding of the input handle returns [`None`].
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or a byte range is invalid.
-    async fn partial_decode_many<'a>(
-        &'a self,
-        decoded_regions: ByteRangeIterator<'a>,
-        options: &CodecOptions,
-    ) -> Result<Option<Vec<ArrayBytesRaw<'a>>>, CodecError>;
-
-    /// Decode all bytes.
-    ///
-    /// Returns [`None`] if partial decoding of the input handle returns [`None`].
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails.
-    async fn decode<'a>(
-        &'a self,
-        options: &CodecOptions,
-    ) -> Result<Option<ArrayBytesRaw<'a>>, CodecError> {
-        self.partial_decode(ByteRange::FromStart(0, None), options)
-            .await
-    }
-
-    /// Returns whether this decoder supports partial decoding.
-    ///
-    /// If this returns `true`, the decoder can efficiently handle partial decoding operations.
-    /// If this returns `false`, partial decoding will fall back to a full decode operation.
-    fn supports_partial_decode(&self) -> bool;
-}
-
-/// Partial array decoder traits.
-pub trait ArrayPartialDecoderTraits: Any + MaybeSend + MaybeSync {
-    /// Return the data type of the partial decoder.
-    fn data_type(&self) -> &DataType;
-
-    /// Returns whether the chunk exists.
-    ///
-    /// # Errors
-    /// Returns [`StorageError`] if a storage operation fails.
-    fn exists(&self) -> Result<bool, StorageError>;
-
-    /// Returns the size of chunk bytes held by the partial decoder.
-    ///
-    /// Intended for use by size-constrained partial decoder caches.
-    fn size_held(&self) -> usize;
-
-    /// Partially decode a chunk.
-    ///
-    /// If the inner `input_handle` is a bytes decoder and partial decoding returns [`None`], then the array subsets have the fill value.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or an array subset is invalid.
-    fn partial_decode(
-        &self,
-        indexer: &dyn Indexer,
-        options: &CodecOptions,
-    ) -> Result<ArrayBytes<'_>, CodecError>;
-
-    /// Partially decode into a preallocated output.
-    ///
-    /// This method is intended for internal use by Array.
-    /// It currently only works for fixed length data types.
-    ///
-    /// The `indexer` shape and dimensionality does not need to match `output_subset`, but the number of elements must match.
-    /// Extracted elements from the `indexer` are written as ordered by the indexer.
-    /// For an [`ArraySubset`], that is C order.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or the number of elements in `indexer` does not match the number of elements in `output_view`,
-    fn partial_decode_into(
-        &self,
-        indexer: &dyn Indexer,
-        output_target: ArrayBytesDecodeIntoTarget<'_>,
-        options: &CodecOptions,
-    ) -> Result<(), CodecError> {
-        if indexer.len() != output_target.num_elements() {
-            return Err(InvalidNumberOfElementsError::new(
-                indexer.len(),
-                output_target.num_elements(),
-            )
-            .into());
-        }
-
-        let decoded_value = self.partial_decode(indexer, options)?;
-        decode_into_array_bytes_target(&decoded_value, output_target)
-    }
-
-    /// Returns whether this decoder supports partial decoding.
-    ///
-    /// If this returns `true`, the decoder can efficiently handle partial decoding operations.
-    /// If this returns `false`, partial decoding will fall back to a full decode operation.
-    fn supports_partial_decode(&self) -> bool;
-}
-
-/// Partial array encoder traits.
-pub trait ArrayPartialEncoderTraits:
-    ArrayPartialDecoderTraits + Any + MaybeSend + MaybeSync
-{
-    /// Return the encoder as an [`Arc<ArrayPartialDecoderTraits>`].
-    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn ArrayPartialDecoderTraits>;
-
-    /// Erase the chunk.
-    ///
-    /// # Errors
-    /// Returns an error if there is an underlying store error.
-    fn erase(&self) -> Result<(), CodecError>;
-
-    /// Partially encode a chunk.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or an array subset is invalid.
-    fn partial_encode(
-        &self,
-        indexer: &dyn Indexer,
-        bytes: &ArrayBytes<'_>,
-        options: &CodecOptions,
-    ) -> Result<(), CodecError>;
-
-    /// Returns whether this encoder supports partial encoding.
-    ///
-    /// If this returns `true`, the encoder can efficiently handle partial encoding operations.
-    /// If this returns `false`, partial encoding will fall back to a full decode and encode operation.
-    fn supports_partial_encode(&self) -> bool;
-}
-
-#[cfg(feature = "async")]
-/// Asynchronous partial array encoder traits.
-#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
-pub trait AsyncArrayPartialEncoderTraits:
-    AsyncArrayPartialDecoderTraits + Any + MaybeSend + MaybeSync
-{
-    /// Return the encoder as an [`Arc<AsyncArrayPartialDecoderTraits>`].
-    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn AsyncArrayPartialDecoderTraits>;
-
-    /// Erase the chunk.
-    ///
-    /// # Errors
-    /// Returns an error if there is an underlying store error.
-    async fn erase(&self) -> Result<(), CodecError>;
-
-    /// Partially encode a chunk.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or an array subset is invalid.
-    async fn partial_encode(
-        &self,
-        indexer: &dyn Indexer,
-        bytes: &ArrayBytes<'_>,
-        options: &CodecOptions,
-    ) -> Result<(), CodecError>;
-
-    /// Returns whether this encoder supports partial encoding.
-    ///
-    /// If this returns `true`, the encoder can efficiently handle partial encoding operations.
-    /// If this returns `false`, partial encoding will fall back to a full decode and encode operation.
-    fn supports_partial_encode(&self) -> bool;
-}
-
-/// Partial bytes encoder traits.
-pub trait BytesPartialEncoderTraits:
-    BytesPartialDecoderTraits + Any + MaybeSend + MaybeSync
-{
-    /// Return the encoder as an [`Arc<BytesPartialDecoderTraits>`].
-    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn BytesPartialDecoderTraits>;
-
-    /// Erase the chunk.
-    ///
-    /// # Errors
-    /// Returns an error if there is an underlying store error.
-    fn erase(&self) -> Result<(), CodecError>;
-
-    /// Partially encode a chunk.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or the byte range is invalid.
-    fn partial_encode(
-        &self,
-        offset: u64,
-        bytes: ArrayBytesRaw<'_>,
-        options: &CodecOptions,
-    ) -> Result<(), CodecError> {
-        self.partial_encode_many(Box::new([(offset, bytes)].into_iter()), options)
-    }
-
-    /// Partially encode a chunk from an [`OffsetBytesIterator`].
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or a byte range is invalid.
-    fn partial_encode_many(
-        &self,
-        offset_values: OffsetBytesIterator<ArrayBytesRaw<'_>>,
-        options: &CodecOptions,
-    ) -> Result<(), CodecError>;
-
-    /// Returns whether this encoder supports partial encoding.
-    ///
-    /// If this returns `true`, the encoder can efficiently handle partial encoding operations.
-    /// If this returns `false`, partial encoding will fall back to a full decode and encode operation.
-    fn supports_partial_encode(&self) -> bool;
-}
-
-#[cfg(feature = "async")]
-/// Asynhronous partial bytes encoder traits.
-#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
-pub trait AsyncBytesPartialEncoderTraits:
-    AsyncBytesPartialDecoderTraits + Any + MaybeSend + MaybeSync
-{
-    /// Return the encoder as an [`Arc<AsyncBytesPartialDecoderTraits>`].
-    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn AsyncBytesPartialDecoderTraits>;
-
-    /// Erase the chunk.
-    ///
-    /// # Errors
-    /// Returns an error if there is an underlying store error.
-    async fn erase(&self) -> Result<(), CodecError>;
-
-    /// Partially encode a chunk.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or the byte range is invalid.
-    async fn partial_encode(
-        &self,
-        offset: u64,
-        bytes: ArrayBytesRaw<'_>,
-        options: &CodecOptions,
-    ) -> Result<(), CodecError> {
-        self.partial_encode_many(Box::new([(offset, bytes)].into_iter()), options)
-            .await
-    }
-
-    /// Partially encode a chunk.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or a byte range is invalid.
-    async fn partial_encode_many<'a>(
-        &'a self,
-        offset_values: OffsetBytesIterator<'a, ArrayBytesRaw<'_>>,
-        options: &CodecOptions,
-    ) -> Result<(), CodecError>;
-
-    /// Returns whether this encoder supports partial encoding.
-    ///
-    /// If this returns `true`, the encoder can efficiently handle partial encoding operations.
-    /// If this returns `false`, partial encoding will fall back to a full decode and encode operation.
-    fn supports_partial_encode(&self) -> bool;
-}
-
-#[cfg(feature = "async")]
-/// Asynchronous partial array decoder traits.
-#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
-pub trait AsyncArrayPartialDecoderTraits: Any + MaybeSend + MaybeSync {
-    /// Return the data type of the partial decoder.
-    fn data_type(&self) -> &DataType;
-
-    /// Returns whether the chunk exists.
-    ///
-    /// # Errors
-    /// Returns [`StorageError`] if a storage operation fails.
-    async fn exists(&self) -> Result<bool, StorageError>;
-
-    /// Returns the size of chunk bytes held by the partial decoder.
-    ///
-    /// Intended for use by size-constrained partial decoder caches.
-    fn size_held(&self) -> usize;
-
-    /// Partially decode a chunk.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails, array subset is invalid, or the array subset shape does not match array view subset shape.
-    async fn partial_decode<'a>(
-        &'a self,
-        indexer: &dyn Indexer,
-        options: &CodecOptions,
-    ) -> Result<ArrayBytes<'a>, CodecError>;
-
-    /// Async variant of [`ArrayPartialDecoderTraits::partial_decode_into`].
-    #[allow(clippy::missing_safety_doc)]
-    async fn partial_decode_into(
-        &self,
-        indexer: &dyn Indexer,
-        output_target: ArrayBytesDecodeIntoTarget<'_>,
-        options: &CodecOptions,
-    ) -> Result<(), CodecError> {
-        if indexer.len() != output_target.num_elements() {
-            return Err(InvalidNumberOfElementsError::new(
-                indexer.len(),
-                output_target.num_elements(),
-            )
-            .into());
-        }
-
-        let decoded_value = self.partial_decode(indexer, options).await?;
-        decode_into_array_bytes_target(&decoded_value, output_target)
-    }
-
-    /// Returns whether this decoder supports partial decoding.
-    ///
-    /// If this returns `true`, the decoder can efficiently handle partial decoding operations.
-    /// If this returns `false`, partial decoding will fall back to a full decode operation.
-    fn supports_partial_decode(&self) -> bool;
-}
-
 /// A [`ReadableStorage`] store value partial decoder.
 pub struct StoragePartialDecoder {
     storage: ReadableStorage,
@@ -938,75 +474,6 @@ impl AsyncBytesPartialDecoderTraits for AsyncStoragePartialDecoder {
 
     fn supports_partial_decode(&self) -> bool {
         self.storage.supports_get_partial()
-    }
-}
-
-impl BytesPartialDecoderTraits for Mutex<Option<Vec<u8>>> {
-    fn exists(&self) -> Result<bool, StorageError> {
-        Ok(self.lock().unwrap().is_some())
-    }
-
-    fn size_held(&self) -> usize {
-        self.lock().unwrap().as_ref().map_or(0, Vec::len)
-    }
-
-    fn partial_decode_many(
-        &self,
-        decoded_regions: ByteRangeIterator,
-        _options: &CodecOptions,
-    ) -> Result<Option<Vec<ArrayBytesRaw<'_>>>, CodecError> {
-        if let Some(input) = self.lock().unwrap().as_ref() {
-            let size = input.len() as u64;
-            let mut outputs = vec![];
-            for byte_range in decoded_regions {
-                if byte_range.end(size) <= size {
-                    outputs.push(Cow::Owned(input[byte_range.to_range_usize(size)].into()));
-                } else {
-                    return Err(InvalidByteRangeError::new(byte_range, size).into());
-                }
-            }
-            Ok(Some(outputs))
-        } else {
-            Ok(None)
-        }
-    }
-
-    fn supports_partial_decode(&self) -> bool {
-        true
-    }
-}
-
-impl BytesPartialEncoderTraits for Mutex<Option<Vec<u8>>> {
-    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn BytesPartialDecoderTraits> {
-        self.clone()
-    }
-
-    fn erase(&self) -> Result<(), CodecError> {
-        *self.lock().unwrap() = None;
-        Ok(())
-    }
-
-    fn partial_encode_many(
-        &self,
-        offset_values: OffsetBytesIterator<ArrayBytesRaw<'_>>,
-        _options: &CodecOptions,
-    ) -> Result<(), CodecError> {
-        let mut v = self.lock().unwrap();
-        let mut output = v.as_ref().cloned().unwrap_or_default();
-
-        for (offset, value) in offset_values {
-            let offset = usize::try_from(offset).unwrap();
-            if output.len() < offset + value.len() {
-                output.resize(offset + value.len(), 0);
-            }
-            output[offset..offset + value.len()].copy_from_slice(&value);
-        }
-        *v = Some(output);
-        Ok(())
-    }
-
-    fn supports_partial_encode(&self) -> bool {
-        true
     }
 }
 
@@ -1149,661 +616,6 @@ impl AsyncBytesPartialEncoderTraits for StoragePartialEncoder<AsyncReadableWrita
 
     fn supports_partial_encode(&self) -> bool {
         self.storage.supports_set_partial()
-    }
-}
-
-/// Traits for array to array codecs.
-#[cfg_attr(
-    all(feature = "async", not(target_arch = "wasm32")),
-    async_trait::async_trait
-)]
-#[cfg_attr(all(feature = "async", target_arch = "wasm32"), async_trait::async_trait(?Send))]
-pub trait UnboundArrayToArrayCodecTraits: CodecTraits + core::fmt::Debug {
-    /// Return a dynamic version of the codec.
-    fn into_dyn(self: Arc<Self>) -> Arc<dyn UnboundArrayToArrayCodecTraits>;
-
-    /// Return a version of this codec reconfigured with the provided codec-specific options.
-    ///
-    /// The default implementation returns the codec unchanged.
-    /// Override this to read your codec's options type from [`CodecSpecificOptions`].
-    ///
-    /// # Errors
-    /// Returns a [`CodecCreateError`] if the codec cannot be reconfigured.
-    #[expect(unused_variables)]
-    fn with_codec_specific_options(
-        self: Arc<Self>,
-        opts: &CodecSpecificOptions,
-    ) -> Result<Arc<dyn UnboundArrayToArrayCodecTraits>, CodecCreateError> {
-        Ok(self.into_dyn())
-    }
-
-    /// Bind this codec to a decoded data type and fill value.
-    ///
-    /// Binding eagerly validates and derives the encoded context.
-    ///
-    /// # Errors
-    /// Returns a [`CodecCreateError`] if the `data_type` or `fill_value` is not supported by this codec.
-    fn with_context(
-        &self,
-        data_type: DataType,
-        fill_value: FillValue,
-    ) -> Result<Arc<dyn ArrayToArrayCodecTraits>, CodecCreateError>;
-}
-
-/// Runtime traits for an array-to-array codec bound to a data type and fill value.
-#[cfg_attr(
-    all(feature = "async", not(target_arch = "wasm32")),
-    async_trait::async_trait
-)]
-#[cfg_attr(all(feature = "async", target_arch = "wasm32"), async_trait::async_trait(?Send))]
-pub trait ArrayToArrayCodecTraits: ArrayCodecTraits + core::fmt::Debug {
-    /// Return a dynamic version of the bound codec.
-    fn into_dyn(self: Arc<Self>) -> Arc<dyn ArrayToArrayCodecTraits>;
-
-    /// Return the encoded data type.
-    fn encoded_data_type(&self) -> &DataType;
-
-    /// Return the encoded fill value.
-    fn encoded_fill_value(&self) -> &FillValue;
-
-    /// Returns the shape of the encoded chunk for a given decoded chunk shape.
-    ///
-    /// # Errors
-    /// Returns a [`CodecError`] if the `decoded_shape` is not supported by this codec.
-    fn encoded_shape(&self, decoded_shape: &[NonZeroU64]) -> Result<ChunkShape, CodecError> {
-        Ok(decoded_shape.to_vec())
-    }
-
-    /// Map a partial decode granularity from the encoded representation to the decoded representation.
-    ///
-    /// # Errors
-    /// Returns a [`CodecError`] if the decoded shape or encoded granularity is not supported by this codec.
-    fn partial_decode_granularity(
-        &self,
-        decoded_shape: &[NonZeroU64],
-        encoded_granularity: &[NonZeroU64],
-    ) -> Result<ChunkShape, CodecError> {
-        let encoded_shape = self.encoded_shape(decoded_shape)?;
-        if encoded_shape == decoded_shape {
-            Ok(encoded_granularity.to_vec())
-        } else {
-            Ok(decoded_shape.to_vec())
-        }
-    }
-
-    /// Encode a chunk.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or `bytes` is incompatible with the decoded representation.
-    fn encode<'a>(
-        &self,
-        bytes: ArrayBytes<'a>,
-        shape: &[NonZeroU64],
-        options: &CodecOptions,
-    ) -> Result<ArrayBytes<'a>, CodecError>;
-
-    /// Decode a chunk.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or the decoded output is incompatible with the decoded representation.
-    fn decode<'a>(
-        &self,
-        bytes: ArrayBytes<'a>,
-        shape: &[NonZeroU64],
-        options: &CodecOptions,
-    ) -> Result<ArrayBytes<'a>, CodecError>;
-
-    /// Initialise a partial decoder.
-    ///
-    /// # Errors
-    /// Returns a [`CodecError`] if initialisation fails.
-    fn partial_decoder(
-        self: Arc<Self>,
-        input_handle: Arc<dyn ArrayPartialDecoderTraits>,
-        shape: &[NonZeroU64],
-        options: &CodecOptions,
-    ) -> Result<Arc<dyn ArrayPartialDecoderTraits>, CodecError> {
-        _ = options;
-        Ok(Arc::new(ArrayToArrayCodecPartialDefault::new(
-            input_handle,
-            shape.to_vec(),
-            self.data_type().clone(),
-            self.fill_value().clone(),
-            self.into_dyn(),
-        )))
-    }
-
-    /// Initialise a partial encoder.
-    ///
-    /// # Errors
-    /// Returns a [`CodecError`] if initialisation fails.
-    fn partial_encoder(
-        self: Arc<Self>,
-        input_output_handle: Arc<dyn ArrayPartialEncoderTraits>,
-        shape: &[NonZeroU64],
-        options: &CodecOptions,
-    ) -> Result<Arc<dyn ArrayPartialEncoderTraits>, CodecError> {
-        _ = options;
-        Ok(Arc::new(ArrayToArrayCodecPartialDefault::new(
-            input_output_handle,
-            shape.to_vec(),
-            self.data_type().clone(),
-            self.fill_value().clone(),
-            self.into_dyn(),
-        )))
-    }
-
-    #[cfg(feature = "async")]
-    /// Initialise an asynchronous partial decoder.
-    ///
-    /// # Errors
-    /// Returns a [`CodecError`] if initialisation fails.
-    async fn async_partial_decoder(
-        self: Arc<Self>,
-        input_handle: Arc<dyn AsyncArrayPartialDecoderTraits>,
-        shape: &[NonZeroU64],
-        options: &CodecOptions,
-    ) -> Result<Arc<dyn AsyncArrayPartialDecoderTraits>, CodecError> {
-        _ = options;
-        Ok(Arc::new(ArrayToArrayCodecPartialDefault::new(
-            input_handle,
-            shape.to_vec(),
-            self.data_type().clone(),
-            self.fill_value().clone(),
-            self.into_dyn(),
-        )))
-    }
-
-    #[cfg(feature = "async")]
-    /// Initialise an asynchronous partial encoder.
-    ///
-    /// # Errors
-    /// Returns a [`CodecError`] if initialisation fails.
-    async fn async_partial_encoder(
-        self: Arc<Self>,
-        input_output_handle: Arc<dyn AsyncArrayPartialEncoderTraits>,
-        shape: &[NonZeroU64],
-        options: &CodecOptions,
-    ) -> Result<Arc<dyn AsyncArrayPartialEncoderTraits>, CodecError> {
-        _ = options;
-        Ok(Arc::new(ArrayToArrayCodecPartialDefault::new(
-            input_output_handle,
-            shape.to_vec(),
-            self.data_type().clone(),
-            self.fill_value().clone(),
-            self.into_dyn(),
-        )))
-    }
-}
-
-/// Traits for array to bytes codecs.
-#[cfg_attr(
-    all(feature = "async", not(target_arch = "wasm32")),
-    async_trait::async_trait
-)]
-#[cfg_attr(all(feature = "async", target_arch = "wasm32"), async_trait::async_trait(?Send))]
-pub trait UnboundArrayToBytesCodecTraits: CodecTraits + core::fmt::Debug {
-    /// Return a dynamic version of the codec.
-    fn into_dyn(self: Arc<Self>) -> Arc<dyn UnboundArrayToBytesCodecTraits>;
-
-    /// Return a version of this codec reconfigured with the provided codec-specific options.
-    ///
-    /// The default implementation returns the codec unchanged.
-    /// Override this to read your codec's options type from [`CodecSpecificOptions`].
-    ///
-    /// # Errors
-    /// Returns a [`CodecCreateError`] if the codec cannot be reconfigured.
-    #[expect(unused_variables)]
-    fn with_codec_specific_options(
-        self: Arc<Self>,
-        opts: &CodecSpecificOptions,
-    ) -> Result<Arc<dyn UnboundArrayToBytesCodecTraits>, CodecCreateError> {
-        Ok(self.into_dyn())
-    }
-
-    /// Bind this codec to a decoded data type and fill value.
-    ///
-    /// # Errors
-    /// Returns a [`CodecCreateError`] if the `data_type` or `fill_value` is not supported by this codec.
-    fn with_context(
-        &self,
-        data_type: DataType,
-        fill_value: FillValue,
-    ) -> Result<Arc<dyn ArrayToBytesCodecTraits>, CodecCreateError>;
-}
-
-/// Runtime traits for an array-to-bytes codec bound to a data type and fill value.
-#[cfg_attr(
-    all(feature = "async", not(target_arch = "wasm32")),
-    async_trait::async_trait
-)]
-#[cfg_attr(all(feature = "async", target_arch = "wasm32"), async_trait::async_trait(?Send))]
-pub trait ArrayToBytesCodecTraits: ArrayCodecTraits + core::fmt::Debug {
-    /// Return a dynamic version of the bound codec.
-    fn into_dyn(self: Arc<Self>) -> Arc<dyn ArrayToBytesCodecTraits>;
-
-    /// Returns the size of the encoded representation given a size of the decoded representation.
-    ///
-    /// # Errors
-    /// Returns a [`CodecError`] if the decoded representation is not supported by this codec.
-    fn encoded_representation(
-        &self,
-        shape: &[NonZeroU64],
-    ) -> Result<BytesRepresentation, CodecError>;
-
-    /// Return the partial decode granularity.
-    ///
-    /// This represents the shape of the smallest subset of a chunk that can be efficiently decoded if the chunk were subdivided into a regular grid.
-    /// For most codecs, this is just the shape of the chunk.
-    /// It is the shape of the sub chunks (inner chunks) for the sharding codec.
-    ///
-    /// # Errors
-    /// Returns a [`CodecError`] if the decoded shape is not supported by this codec.
-    fn partial_decode_granularity(
-        &self,
-        decoded_shape: &[NonZeroU64],
-    ) -> Result<ChunkShape, CodecError> {
-        Ok(decoded_shape.to_vec())
-    }
-
-    /// Encode a chunk.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or `bytes` is incompatible with the decoded representation.
-    fn encode<'a>(
-        &self,
-        bytes: ArrayBytes<'a>,
-        shape: &[NonZeroU64],
-        options: &CodecOptions,
-    ) -> Result<ArrayBytesRaw<'a>, CodecError>;
-
-    /// Decode a chunk.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or the decoded output is incompatible with the decoded representation.
-    fn decode<'a>(
-        &self,
-        bytes: ArrayBytesRaw<'a>,
-        shape: &[NonZeroU64],
-        options: &CodecOptions,
-    ) -> Result<ArrayBytes<'a>, CodecError>;
-
-    /// Compact a chunk to remove any extraneous data.
-    ///
-    /// The default implementation returns the input `bytes` unchanged.
-    ///
-    /// Returns `Ok(None)` if no compaction was performed.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or `bytes` is incompatible with the decoded representation.
-    #[allow(unused_variables)]
-    fn compact<'a>(
-        &self,
-        bytes: ArrayBytesRaw<'a>,
-        shape: &[NonZeroU64],
-        options: &CodecOptions,
-    ) -> Result<Option<ArrayBytesRaw<'a>>, CodecError> {
-        Ok(None)
-    }
-
-    /// Decode into a subset of a preallocated output.
-    ///
-    /// This method is intended for internal use by Array.
-    /// It works for fixed length data types and optional data types.
-    ///
-    /// The decoded representation shape and dimensionality does not need to match the output target, but the number of elements must match.
-    /// Chunk elements are written to the subset of the output in C order.
-    ///
-    /// For optional data types, provide an `ArrayBytesDecodeIntoTarget` with a `mask` set to `Some`.
-    /// For non-optional data types, convert a fixed view to target using `.into()` or create with `mask: None`.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails or the number of elements in the decoded representation does not match the number of elements in the output target.
-    fn decode_into(
-        &self,
-        bytes: ArrayBytesRaw<'_>,
-        shape: &[NonZeroU64],
-        output_target: ArrayBytesDecodeIntoTarget<'_>,
-        options: &CodecOptions,
-    ) -> Result<(), CodecError> {
-        let bytes = self.decode(bytes, shape, options)?;
-        decode_into_array_bytes_target(&bytes, output_target)
-    }
-
-    /// Initialise a partial decoder.
-    /// # Errors
-    /// Returns a [`CodecError`] if initialisation fails.
-    fn partial_decoder(
-        self: Arc<Self>,
-        input_handle: Arc<dyn BytesPartialDecoderTraits>,
-        shape: &[NonZeroU64],
-        options: &CodecOptions,
-    ) -> Result<Arc<dyn ArrayPartialDecoderTraits>, CodecError> {
-        _ = options;
-        Ok(Arc::new(ArrayToBytesCodecPartialDefault::new(
-            input_handle,
-            shape.to_vec(),
-            self.data_type().clone(),
-            self.fill_value().clone(),
-            self.into_dyn(),
-        )))
-    }
-
-    /// Initialise a partial encoder.
-    ///
-    /// # Errors
-    /// Returns a [`CodecError`] if initialisation fails.
-    fn partial_encoder(
-        self: Arc<Self>,
-        input_output_handle: Arc<dyn BytesPartialEncoderTraits>,
-        shape: &[NonZeroU64],
-        options: &CodecOptions,
-    ) -> Result<Arc<dyn ArrayPartialEncoderTraits>, CodecError> {
-        _ = options;
-        Ok(Arc::new(ArrayToBytesCodecPartialDefault::new(
-            input_output_handle,
-            shape.to_vec(),
-            self.data_type().clone(),
-            self.fill_value().clone(),
-            self.into_dyn(),
-        )))
-    }
-
-    #[cfg(feature = "async")]
-    /// Initialise an asynchronous partial decoder.
-    ///
-    /// # Errors
-    /// Returns a [`CodecError`] if initialisation fails.
-    async fn async_partial_decoder(
-        self: Arc<Self>,
-        input_handle: Arc<dyn AsyncBytesPartialDecoderTraits>,
-        shape: &[NonZeroU64],
-        options: &CodecOptions,
-    ) -> Result<Arc<dyn AsyncArrayPartialDecoderTraits>, CodecError> {
-        _ = options;
-        Ok(Arc::new(ArrayToBytesCodecPartialDefault::new(
-            input_handle,
-            shape.to_vec(),
-            self.data_type().clone(),
-            self.fill_value().clone(),
-            self.into_dyn(),
-        )))
-    }
-
-    #[cfg(feature = "async")]
-    /// Initialise an asynchronous partial encoder.
-    ///
-    /// # Errors
-    /// Returns a [`CodecError`] if initialisation fails.
-    async fn async_partial_encoder(
-        self: Arc<Self>,
-        input_output_handle: Arc<dyn AsyncBytesPartialEncoderTraits>,
-        shape: &[NonZeroU64],
-        options: &CodecOptions,
-    ) -> Result<Arc<dyn AsyncArrayPartialEncoderTraits>, CodecError> {
-        _ = options;
-        Ok(Arc::new(ArrayToBytesCodecPartialDefault::new(
-            input_output_handle,
-            shape.to_vec(),
-            self.data_type().clone(),
-            self.fill_value().clone(),
-            self.into_dyn(),
-        )))
-    }
-}
-
-/// Traits for bytes to bytes codecs.
-#[cfg_attr(
-    all(feature = "async", not(target_arch = "wasm32")),
-    async_trait::async_trait
-)]
-#[cfg_attr(all(feature = "async", target_arch = "wasm32"), async_trait::async_trait(?Send))]
-pub trait BytesToBytesCodecTraits: CodecTraits + core::fmt::Debug {
-    /// Return a dynamic version of the codec.
-    fn into_dyn(self: Arc<Self>) -> Arc<dyn BytesToBytesCodecTraits>;
-
-    /// Return a version of this codec reconfigured with the provided codec-specific options.
-    ///
-    /// The default implementation returns the codec unchanged.
-    /// Override this to read your codec's options type from [`CodecSpecificOptions`].
-    #[expect(unused_variables)]
-    fn with_codec_specific_options(
-        self: Arc<Self>,
-        opts: &CodecSpecificOptions,
-    ) -> Result<Arc<dyn BytesToBytesCodecTraits>, CodecCreateError> {
-        Ok(self.into_dyn())
-    }
-
-    /// Return the maximum internal concurrency supported for the requested decoded representation.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if the decoded representation is not valid for the codec.
-    fn recommended_concurrency(
-        &self,
-        decoded_representation: &BytesRepresentation,
-    ) -> Result<RecommendedConcurrency, CodecError>;
-
-    /// Returns the size of the encoded representation given a size of the decoded representation.
-    fn encoded_representation(
-        &self,
-        decoded_representation: &BytesRepresentation,
-    ) -> BytesRepresentation;
-
-    /// Encode chunk bytes.
-    ///
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails.
-    fn encode<'a>(
-        &self,
-        decoded_value: ArrayBytesRaw<'a>,
-        options: &CodecOptions,
-    ) -> Result<ArrayBytesRaw<'a>, CodecError>;
-
-    /// Decode chunk bytes.
-    //
-    /// # Errors
-    /// Returns [`CodecError`] if a codec fails.
-    fn decode<'a>(
-        &self,
-        encoded_value: ArrayBytesRaw<'a>,
-        decoded_representation: &BytesRepresentation,
-        options: &CodecOptions,
-    ) -> Result<ArrayBytesRaw<'a>, CodecError>;
-
-    /// Initialises a partial decoder.
-    ///
-    /// The default implementation decodes the entire chunk.
-    ///
-    /// # Errors
-    /// Returns a [`CodecError`] if initialisation fails.
-    #[allow(unused_variables)]
-    fn partial_decoder(
-        self: Arc<Self>,
-        input_handle: Arc<dyn BytesPartialDecoderTraits>,
-        decoded_representation: &BytesRepresentation,
-        options: &CodecOptions,
-    ) -> Result<Arc<dyn BytesPartialDecoderTraits>, CodecError> {
-        Ok(Arc::new(BytesToBytesCodecPartialDefault::new_bytes(
-            input_handle,
-            *decoded_representation,
-            self.into_dyn(),
-        )))
-    }
-
-    /// Initialise a partial encoder.
-    ///
-    /// The default implementation reencodes the entire chunk.
-    ///
-    /// # Errors
-    /// Returns a [`CodecError`] if initialisation fails.
-    #[allow(unused_variables)]
-    fn partial_encoder(
-        self: Arc<Self>,
-        input_output_handle: Arc<dyn BytesPartialEncoderTraits>,
-        decoded_representation: &BytesRepresentation,
-        options: &CodecOptions,
-    ) -> Result<Arc<dyn BytesPartialEncoderTraits>, CodecError> {
-        Ok(Arc::new(BytesToBytesCodecPartialDefault::new_bytes(
-            input_output_handle,
-            *decoded_representation,
-            self.into_dyn(),
-        )))
-    }
-
-    #[cfg(feature = "async")]
-    /// Initialises an asynchronous partial decoder.
-    ///
-    /// The default implementation decodes the entire chunk.
-    ///
-    /// # Errors
-    /// Returns a [`CodecError`] if initialisation fails.
-    #[allow(unused_variables)]
-    async fn async_partial_decoder(
-        self: Arc<Self>,
-        input_handle: Arc<dyn AsyncBytesPartialDecoderTraits>,
-        decoded_representation: &BytesRepresentation,
-        options: &CodecOptions,
-    ) -> Result<Arc<dyn AsyncBytesPartialDecoderTraits>, CodecError> {
-        Ok(Arc::new(BytesToBytesCodecPartialDefault::new_bytes(
-            input_handle,
-            *decoded_representation,
-            self.into_dyn(),
-        )))
-    }
-
-    #[cfg(feature = "async")]
-    /// Initialise an asynchronous partial encoder.
-    ///
-    /// The default implementation reencodes the entire chunk.
-    ///
-    /// # Errors
-    /// Returns a [`CodecError`] if initialisation fails.
-    #[allow(unused_variables)]
-    async fn async_partial_encoder(
-        self: Arc<Self>,
-        input_output_handle: Arc<dyn AsyncBytesPartialEncoderTraits>,
-        decoded_representation: &BytesRepresentation,
-        options: &CodecOptions,
-    ) -> Result<Arc<dyn AsyncBytesPartialEncoderTraits>, CodecError> {
-        Ok(Arc::new(BytesToBytesCodecPartialDefault::new_bytes(
-            input_output_handle,
-            *decoded_representation,
-            self.into_dyn(),
-        )))
-    }
-}
-
-impl BytesPartialDecoderTraits for Cow<'static, [u8]> {
-    fn exists(&self) -> Result<bool, StorageError> {
-        Ok(true)
-    }
-
-    fn size_held(&self) -> usize {
-        self.as_ref().len()
-    }
-
-    fn partial_decode_many(
-        &self,
-        decoded_regions: ByteRangeIterator,
-        _parallel: &CodecOptions,
-    ) -> Result<Option<Vec<ArrayBytesRaw<'_>>>, CodecError> {
-        Ok(Some(
-            extract_byte_ranges(self, decoded_regions)?
-                .into_iter()
-                .map(Cow::Owned)
-                .collect(),
-        ))
-    }
-
-    fn supports_partial_decode(&self) -> bool {
-        true
-    }
-}
-
-impl BytesPartialDecoderTraits for Vec<u8> {
-    fn exists(&self) -> Result<bool, StorageError> {
-        Ok(true)
-    }
-
-    fn size_held(&self) -> usize {
-        self.len()
-    }
-
-    fn partial_decode_many(
-        &self,
-        decoded_regions: ByteRangeIterator,
-        _parallel: &CodecOptions,
-    ) -> Result<Option<Vec<ArrayBytesRaw<'_>>>, CodecError> {
-        Ok(Some(
-            extract_byte_ranges(self, decoded_regions)?
-                .into_iter()
-                .map(Cow::Owned)
-                .collect(),
-        ))
-    }
-
-    fn supports_partial_decode(&self) -> bool {
-        true
-    }
-}
-
-#[cfg(feature = "async")]
-#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
-impl AsyncBytesPartialDecoderTraits for Cow<'static, [u8]> {
-    async fn exists(&self) -> Result<bool, StorageError> {
-        Ok(true)
-    }
-
-    fn size_held(&self) -> usize {
-        self.as_ref().len()
-    }
-
-    async fn partial_decode_many<'a>(
-        &'a self,
-        decoded_regions: ByteRangeIterator<'a>,
-        _parallel: &CodecOptions,
-    ) -> Result<Option<Vec<ArrayBytesRaw<'a>>>, CodecError> {
-        Ok(Some(
-            extract_byte_ranges(self, decoded_regions)?
-                .into_iter()
-                .map(Cow::Owned)
-                .collect(),
-        ))
-    }
-
-    fn supports_partial_decode(&self) -> bool {
-        true
-    }
-}
-
-#[cfg(feature = "async")]
-#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
-impl AsyncBytesPartialDecoderTraits for Vec<u8> {
-    async fn exists(&self) -> Result<bool, StorageError> {
-        Ok(true)
-    }
-
-    fn size_held(&self) -> usize {
-        self.len()
-    }
-
-    async fn partial_decode_many<'a>(
-        &'a self,
-        decoded_regions: ByteRangeIterator<'a>,
-        _parallel: &CodecOptions,
-    ) -> Result<Option<Vec<ArrayBytesRaw<'a>>>, CodecError> {
-        Ok(Some(
-            extract_byte_ranges(self, decoded_regions)?
-                .into_iter()
-                .map(Cow::Owned)
-                .collect(),
-        ))
-    }
-
-    fn supports_partial_decode(&self) -> bool {
-        true
     }
 }
